+++ sdk/advisor/arm-advisor/CHANGELOG.md
@@ -1,15 +1,40 @@
 # Release History
+    
+## 4.0.0 (2023-01-19)
+    
+**Features**
 
-## 3.1.1 (Unreleased)
+  - Added operation group AdvisorScores
+  - Added Interface AdvisorScoreEntity
+  - Added Interface AdvisorScoreEntityProperties
+  - Added Interface AdvisorScoresGetOptionalParams
+  - Added Interface AdvisorScoresListOptionalParams
+  - Added Interface PathsW6Fm20SubscriptionsSubscriptionidProvidersMicrosoftAdvisorAdvisorscoreGetResponses200ContentApplicationJsonSchema
+  - Added Interface PredictionRequest
+  - Added Interface PredictionResponse
+  - Added Interface PredictOptionalParams
+  - Added Interface ScoreEntity
+  - Added Interface TimeSeriesEntityItem
+  - Added Type Alias AdvisorScoresGetResponse
+  - Added Type Alias AdvisorScoresListResponse
+  - Added Type Alias Duration
+  - Added Type Alias PredictionType
+  - Added Type Alias PredictResponse
+  - Added Type Alias TimeSeriesEntityItemAggregationLevel
+  - Interface ConfigData has a new optional parameter duration
+  - Added Enum KnownDuration
+  - Added Enum KnownPredictionType
+  - Added Enum KnownTimeSeriesEntityItemAggregationLevel
 
-### Features Added
+**Breaking Changes**
 
-### Breaking Changes
-
-### Bugs Fixed
-
-### Other Changes
-
+  - Interface RecommendationsListNextOptionalParams no longer has parameter filter
+  - Interface RecommendationsListNextOptionalParams no longer has parameter skipToken
+  - Interface RecommendationsListNextOptionalParams no longer has parameter top
+  - Interface SuppressionsListNextOptionalParams no longer has parameter skipToken
+  - Interface SuppressionsListNextOptionalParams no longer has parameter top
+    
+    
 ## 3.1.0 (2022-11-14)
     
 **Features**